modelserver/requirements.txt
tensorflow-cpu>=2.12
numpy>=1.26
redis>=5.0
